Revert "[WEB-4356] `gatsby build` optimizations"
